sokoban v0.2: removed lame #r comment prefix
At development time, the Vim syntax file for sed got lost with this
script, and prefixing all comments with #r instead # solved the problem.

After that I've discovered that the culprit was some strange char in
the code. I've removed that char, but forgot to remove the silly #r
prefix. Now it's done.